Module Tar_lwt_unix

Lwt_unix I/O for tar-formatted data

val really_read : Lwt_unix.file_descr -> Cstruct.t -> unit Lwt.t

really_read fd buf fills buf with data from fd or fails with End_of_file

val really_write : Lwt_unix.file_descr -> Cstruct.t -> unit Lwt.t

really_write fd buf writes the full contents of buf to fd or fails with End_of_file

module Header : sig ... end

module Archive : sig ... end