Module Tar_unix

Unix I/O for tar-formatted data

val really_read : Unix.file_descr -> Cstruct.t -> unit

really_read fd buf fills buf with data from fd or raises End_of_file

val really_write : Unix.file_descr -> Cstruct.t -> unit

really_write fd buf writes the full contents of buf to fd or raises End_of_file

module Header : sig ... end

val write_block : ?⁠level:Header.compatibility -> Header.t -> (Unix.file_descr -> unit) -> Unix.file_descr -> unit

Write hdr, then call write_body fd to write the body, then zero-pads so the stream is positioned for the next block.

val write_end : Unix.file_descr -> unit

Write a stream terminator to fd

module Archive : sig ... end