Module Tar_unix.Archive

val with_next_file : Unix.file_descr -> (Unix.file_descr -> Tar.Header.t -> 'a) -> 'a

Read the next header, apply the function 'f' to the fd and the header. The function should leave the fd positioned immediately after the datablock. Finally the function skips past the zero padding to the next header.

val list : ?⁠level:Tar.Header.compatibility -> Unix.file_descr -> Tar.Header.t list

List the contents of a tar.

val extract : (string -> string) -> Unix.file_descr -> unit

extract dest extract the contents of a tar. Apply dest on each source filename to change the destination filename. It only supports extracting regular files from the top-level of the archive.

val transform : ?⁠level:Tar.Header.compatibility -> (Tar.Header.t -> Tar.Header.t) -> Unix.file_descr -> Unix.file_descr -> unit

transform f in_fd out_fd applies f to the header of each file in the tar inputted in in_fd, and writes the resulting headers to out_fd preserving the content and structure of the archive.

val create : string list -> Unix.file_descr -> unit

Create a tar on file descriptor fd from the filename list 'files'. It only supports regular files.

val copy_n : Unix.file_descr -> Unix.file_descr -> int64 -> unit

copy_n ifd odf n copies exactly n bytes from ifd to ofd.

val multicast_n : ?⁠buffer_size:int -> Unix.file_descr -> Unix.file_descr list -> int64 -> unit

multicast_n ?buffer_size ifd ofds n copies exactly n bytes from ifd to all ofds.

val skip : Unix.file_descr -> int -> unit

skip fd n reads and throws away n bytes from fd.